getCountryList common method.
