+++ src/utils/helper.ts
@@ -1,5 +1,7 @@
+import { RequestContextType } from "@providers/request-provider";
 import { CoreModule, getCoreModuleRoute } from "lib/router";
 import { useNavigate } from "react-router-dom";
+import { countryListStorageKey } from "./constants";
 
 export const useCoreModuleNavigation = () => {
   const navigate = useNavigate();
@@ -14,4 +16,19 @@ export const useCoreModuleNavigation = () => {
   };
 
   return handleNavigation;
+};
+
+export const getCountryList = async (context: RequestContextType) => {
+  const countries = localStorage.getItem(countryListStorageKey);
+  if (countries) {
+    return JSON.parse(countries) as Record<string, string>;
+  } else {
+    try {
+      const { data } = await context.client.api.countriesList();
+      localStorage.setItem(countryListStorageKey, JSON.stringify(data));
+      return data;
+    } catch (e) {
+      return null;
+    }
+  }
 };
